requirements.txt
Flask==2.2.2
pwntools==4.9.0
pycryptodome==3.17
requests==2.28.2
